requirements.txt
asgiref==3.5.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
Django==4.0.3
django-filter==21.1
django-widget-tweaks==1.4.12
djangorestframework==3.13.1
hashids==1.3.1
idna==3.3
inflection==0.5.1
Pillow==9.0.1
pytz==2022.1
regex==2022.3.15
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.9
